# Reduce SQL sanitizer allocations 