https://github.com/jackc/pgx/issues/2124

Result:
<img width="863" alt="Pasted Graphic 1" src="https://github.com/user-attachments/assets/c3919d9f-47d9-4bd0-9207-0bad21cc3924">

Main optimizations:
- extensive usage of `sync.Pool` for byte buffers, lexers and parsed query structs
- append-style string formatters for `int64`, `float64` and `time.Time` + `bytes.Buffer.AvailableBuffer`
- rework of `QuoteString` and `QuoteBytes` to append-style (with tests for backwards compatibility) 

Misc changes:
- benchmarks for `Query.Sanitize` and `SanitizeSQL` functions
- a tiny script for generation of benchmark reports for selected commits and diff (using [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat))
- fuzzing of `QuoteString` and `QuoteBytes` (I did'n find any problems for 1h of fuzzing, but you can never be sure for 100%)

Since optimization is an extremely hard problem, I think it's worth checking some more benchmarks.  

I would be very grateful for your opinion on this and recommendations/advice, @jackc @vtolstov