@jackc 2024-10-05
LGTM. But this is obviously a very security critical part of the code, so I'd like if we can get some more eyes on this before merging.