@ninedraft 2024-12-09
@sean I can incorporate your suggestions into this PR if you don't mind